+++ lib/routing/router.dart
@@ -5,6 +5,7 @@ import 'package:hoplixi/core/constants/main_constants.dart';
 import 'package:hoplixi/core/logger/index.dart';
 import 'package:hoplixi/features/cloud_sync/auth/models/auth_flow_status.dart';
 import 'package:hoplixi/features/cloud_sync/auth/providers/auth_flow_provider.dart';
+import 'package:hoplixi/features/cloud_sync/snapshot_sync/providers/current_store_sync_provider.dart';
 import 'package:hoplixi/features/local_send/providers/transfer_provider.dart';
 import 'package:hoplixi/features/setup/providers/setup_completed_provider.dart';
 import 'package:hoplixi/global_key.dart';
@@ -61,6 +62,7 @@ final routerProvider = Provider<GoRouter>((ref) {
         return AppRoutesPaths.home;
       }
       final authFlowState = ref.read(authFlowProvider);
+      final closeStoreSyncStatus = ref.read(closeStoreSyncStatusProvider);
       if (authFlowState.status == AuthFlowStatus.inProgress &&
           currentPath != AppRoutesPaths.cloudSyncAuthProgress) {
         return AppRoutesPaths.cloudSyncAuthProgress;
@@ -72,7 +74,7 @@ final routerProvider = Provider<GoRouter>((ref) {
       if (dbStateAsync.hasValue) {
         final dbState = dbStateAsync.value!;
 
-        if (dbState.isClosingSync) {
+        if (dbState.isClosingSync || closeStoreSyncStatus != null) {
           if (currentPath != AppRoutesPaths.closeStoreSync) {
             if (UniversalPlatform.isDesktop) {
               await WindowManager.instance.setSize(